refactor: 도메인 로직은 domain/, ui 관련 로직은 view/ 로 분리
